requirements.txt
nextcord
psycopg2==2.9.5
python-dotenv==0.21.0
google-api-python-client==2.65.0
oauth2client==4.1.3
httplib2==0.21.0
